Tours: Fix QImage.scaled() with PyQt5 >= 5.15
It looks like PyQt5 >= 5.15 doesn’t accept float arguments in
QImage.scaled(). PyQt5 5.14.1 issues the following warning:

<stdin>:1: DeprecationWarning: an integer is required (got type float).
Implicit conversion to integers using __int__ is deprecated, and may be
removed in a future version of Python.

+++ spyder/plugins/tours/widgets.py
@@ -1069,8 +1069,8 @@ class OpenTourDialog(QDialog):
         image_path = get_image_path(icon_filename)
         image = QPixmap(image_path)
         image_label = QLabel()
-        image_height = image.height() * DialogStyle.IconScaleFactor
-        image_width = image.width() * DialogStyle.IconScaleFactor
+        image_height = int(image.height() * DialogStyle.IconScaleFactor)
+        image_width = int(image.width() * DialogStyle.IconScaleFactor)
         image = image.scaled(image_width, image_height, Qt.KeepAspectRatio,
                              Qt.SmoothTransformation)
         image_label.setPixmap(image)
